RHOAIENG-79882: Move getModelPathFromUri, getPVCNameFromURI, isPVCUri to model-serving shared API
