+++ frontend/src/pages/modelServing/screens/projects/InferenceServiceModal/ConnectionSection.tsx
@@ -14,6 +14,7 @@ import type { PersistentVolumeClaimKind } from '@odh-dashboard/k8s-core';
 import { DashboardPopupIconButton } from '@odh-dashboard/ui-core';
 import { getResourceNameFromK8sResource } from '@odh-dashboard/k8s-core';
 import type { UpdateObjectAtPropAndValue } from '@odh-dashboard/ui-core';
+import { getPVCNameFromURI, isPVCUri } from '@odh-dashboard/model-serving/shared';
 import type { LabeledConnection } from '@odh-dashboard/model-serving/shared';
 import {
   Connection,
@@ -38,12 +39,7 @@ import {
   useNewConnectionField,
   UseNewConnectionFieldData,
 } from '#~/concepts/connectionTypes/NewConnectionField';
-import {
-  getPVCFromURI,
-  getPVCNameFromURI,
-  isModelPathValid,
-  isPVCUri,
-} from '#~/pages/modelServing/screens/projects/utils';
+import { getPVCFromURI, isModelPathValid } from '#~/pages/modelServing/screens/projects/utils';
 import { AccessTypes } from '#~/pages/projects/dataConnections/const';
 import ConnectionS3FolderPathField from './ConnectionS3FolderPathField';
 import ConnectionOciPathField from './ConnectionOciPathField';
