+++ frontend/src/pages/projects/useInferenceServicesForConnection.ts
@@ -1,9 +1,8 @@
 import * as React from 'react';
 import { MetadataAnnotation, type PersistentVolumeClaimKind } from '@odh-dashboard/k8s-core';
-import { InferenceServiceKind } from '@odh-dashboard/model-serving/shared';
+import { InferenceServiceKind, getPVCNameFromURI } from '@odh-dashboard/model-serving/shared';
 import { Connection } from '#~/concepts/connectionTypes/types';
 import { ProjectDetailsContext } from '#~/pages/projects/ProjectDetailsContext';
-import { getPVCNameFromURI } from '#~/pages/modelServing/screens/projects/utils';
 
 export const useInferenceServicesForConnection = (
   connection?: Connection | PersistentVolumeClaimKind,
