+++ packages/model-serving/src/components/deploymentWizard/fields/modelLocationFields/PVCInputField.tsx
@@ -12,8 +12,8 @@ import {
   StackItem,
 } from '@patternfly/react-core';
 import type { PersistentVolumeClaimKind } from '@odh-dashboard/k8s-core';
-import { getModelPathFromUri } from '@odh-dashboard/internal/pages/modelServing/screens/projects/utils';
 import { getModelServingPVCAnnotations } from '@odh-dashboard/model-serving/shared';
+import { getModelPathFromUri } from '../../../../shared/utils/pvcUtils';
 
 type PVCInputFieldProps = {
   selectedPVC?: PersistentVolumeClaimKind;
