+++ packages/model-serving/src/shared/utils/pvcUtils.ts
@@ -8,3 +8,33 @@ export const getModelServingPVCAnnotations = (
 
   return { modelName, modelPath };
 };
+
+export const getPVCNameFromURI = (uri: string): string => {
+  try {
+    const url = new URL(uri);
+    if (url.protocol !== 'pvc:') {
+      return '';
+    }
+    return url.hostname;
+  } catch {
+    return '';
+  }
+};
+
+export const isPVCUri = (uri: string): boolean => {
+  try {
+    const url = new URL(uri);
+    return url.protocol === 'pvc:';
+  } catch {
+    return false;
+  }
+};
+
+export const getModelPathFromUri = (uri: string): string => {
+  try {
+    const url = new URL(uri);
+    return url.pathname.replace(/^\//, '');
+  } catch {
+    return '';
+  }
+};
